Mock quote tidy
