+++ src/lib.rs
@@ -1,5 +1,7 @@
 #![no_std]
 mod error;
+#[cfg(feature = "mock")]
+mod mock;
 mod take_n;
 
 pub use error::QuoteParseError;
@@ -15,7 +17,6 @@ use nom::{
     sequence::tuple,
     IResult,
 };
-use p256::ecdsa::{signature::SignerMut, SigningKey};
 pub use p256::ecdsa::{signature::Verifier, Signature, VerifyingKey};
 
 const QUOTE_HEADER_LENGTH: usize = 48;
@@ -59,7 +60,7 @@ impl Quote {
 
         // Attestation key
         let (input, attestation_key) = take(64u8)(input)?;
-        let attestation_key = [&[04], attestation_key].concat(); // 0x04 means uncompressed
+        let attestation_key = [&[4], attestation_key].concat(); // 0x04 means uncompressed
         let attestation_key = VerifyingKey::from_sec1_bytes(&attestation_key)?;
 
         // Verify signature
@@ -90,52 +91,6 @@ impl Quote {
     /// Returns the build-time measurement register
     pub fn mrtd(&self) -> [u8; 48] {
         self.body.mrtd
-    }
-
-    /// Create a mock quote
-    pub fn mock(mut signing_key: SigningKey, reportdata: [u8; 64]) -> Self {
-        let header = QuoteHeader {
-            version: 4,
-            attestation_key_type: AttestionKeyType::ECDSA256WithP256,
-            tee_type: TEEType::TDX,
-            reserved1: Default::default(),
-            reserved2: Default::default(),
-            qe_vendor_id: Default::default(), // Could use Uuid crate
-            user_data: Default::default(),
-        };
-        let body = QuoteBody {
-            tdx_version: TDXVersion::One,
-            tee_tcb_svn: Default::default(),
-            mrseam: [0; 48],
-            mrsignerseam: [0; 48],
-            seamattributes: Default::default(),
-            tdattributes: Default::default(),
-            xfam: Default::default(),
-            mrtd: [0; 48],
-            mrconfigid: [0; 48],
-            mrowner: [0; 48],
-            mrownerconfig: [0; 48],
-            rtmr0: [0; 48],
-            rtmr1: [0; 48],
-            rtmr2: [0; 48],
-            rtmr3: [0; 48],
-            reportdata,
-            tee_tcb_svn_2: None,
-            mrservicetd: None,
-        };
-        // Serialize header and body to get message to sign
-        let mut message = [0; QUOTE_HEADER_LENGTH + V4_QUOTE_BODY_LENGTH];
-        message[..QUOTE_HEADER_LENGTH].copy_from_slice(&quote_header_serializer(&header));
-        message[QUOTE_HEADER_LENGTH..].copy_from_slice(&quote_body_v4_serializer(&body));
-        let signature = signing_key.sign(&message);
-
-        Quote {
-            header,
-            body,
-            attestation_key: VerifyingKey::from(&signing_key),
-            signature,
-            certification_data: CertificationData::QeReportCertificationData(Default::default()),
-        }
     }
 }
 
@@ -296,27 +251,6 @@ fn quote_header_parser(input: &[u8]) -> IResult<&[u8], QuoteHeader> {
     )(input)
 }
 
-pub fn quote_header_serializer(input: &QuoteHeader) -> [u8; QUOTE_HEADER_LENGTH] {
-    let mut output = [1; QUOTE_HEADER_LENGTH];
-    let version = input.version.to_le_bytes();
-    output[..2].copy_from_slice(&version);
-
-    let attestation_key_type = input.attestation_key_type.clone() as u16;
-    let attestation_key_type = attestation_key_type.to_le_bytes();
-    output[2..4].copy_from_slice(&attestation_key_type);
-
-    let tee_type = input.tee_type.clone() as u32;
-    let tee_type = tee_type.to_le_bytes();
-    output[4..8].copy_from_slice(&tee_type);
-
-    output[8..10].copy_from_slice(&input.reserved1);
-    output[10..12].copy_from_slice(&input.reserved2);
-    output[12..28].copy_from_slice(&input.qe_vendor_id);
-    output[28..48].copy_from_slice(&input.user_data);
-
-    output
-}
-
 /// Parser for a quote body
 fn body_parser(input: &[u8], version: u16) -> IResult<&[u8], QuoteBody> {
     let (input, tdx_version) = match version {
@@ -352,27 +286,7 @@ fn body_parser(input: &[u8], version: u16) -> IResult<&[u8], QuoteBody> {
     } else {
         input
     };
-    return Ok((input, body));
-}
-
-pub fn quote_body_v4_serializer(input: &QuoteBody) -> [u8; V4_QUOTE_BODY_LENGTH] {
-    let mut output = [1; V4_QUOTE_BODY_LENGTH];
-    output[..16].copy_from_slice(&input.tee_tcb_svn);
-    output[16..64].copy_from_slice(&input.mrseam);
-    output[64..112].copy_from_slice(&input.mrsignerseam);
-    output[112..120].copy_from_slice(&input.seamattributes);
-    output[120..128].copy_from_slice(&input.tdattributes);
-    output[128..136].copy_from_slice(&input.xfam);
-    output[136..184].copy_from_slice(&input.mrtd);
-    output[184..232].copy_from_slice(&input.mrconfigid);
-    output[232..280].copy_from_slice(&input.mrowner);
-    output[280..328].copy_from_slice(&input.mrownerconfig);
-    output[328..376].copy_from_slice(&input.rtmr0);
-    output[376..424].copy_from_slice(&input.rtmr1);
-    output[424..472].copy_from_slice(&input.rtmr2);
-    output[472..520].copy_from_slice(&input.rtmr3);
-    output[520..].copy_from_slice(&input.reportdata);
-    output
+    Ok((input, body))
 }
 
 /// Parser for a quote body - omitting optional extra fields for TDX 1.5
